+++ src/com/koxudaxi/ruff/RuffExternalAnnotator.kt
@@ -38,6 +38,7 @@ class RuffExternalAnnotator :
         if (file !is PyFile) return null
         val project = file.project
         val config = RuffConfigService.getInstance(project)
+        if (config.enableLsp) return null
         if (config.useRuffLsp && config.ruffLspExecutablePath is String) return null
         if (config.useRuffServer && RuffCacheService.hasLsp(project) == true) return null
         if (!file.isApplicableTo) return null